Move WhitelistStore to its own file
